Simplify GDR
Guaranteed damage reduction is intended to smooth out some of the
nastiness in Crawl combat. It guarantees that your AC will never
roll below some % of its max value when defending against melee attacks
and mundane projectiles (not against spells, etc).

This is a perfectly fine idea, but it's implemented bafflingly: it uses
your base AC OR a hodge-podge of special cases based on your species
and/or form, meaning that it's both spoilery and weird. Some AC is
effectively worth more than others. Not great!

Instead, simplify it to be based purely on the player's current AC.
(Specifically, AC^(1/4) * 17.)

Previous example values:

Robe:                 0%
Scale mail (ie Fi):  28%
Plate:               39%

New values:

AC 6 (ie Fi):        28%
AC 20:		     38%
AC 50:               47%

This is a buff to some things and a nerf to other things. It's all over
the place. Chaos reigns.